creator-suite-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.0
aiofiles==23.2.0
redis==5.0.1
celery==5.3.4
razorpay==1.4.1
discord.py==2.3.2
python-telegram-bot==20.7
httpx==0.25.2
Pillow==10.1.0
opencv-python==4.8.1.78
moviepy==1.0.3
requests==2.31.0
pyotp==2.9.0
email-validator==2.1.0
jinja2==3.1.2
python-jose==3.3.0
cryptography>=40.0.0
user-agents==2.2.0
geoip2==4.8.0
